+++ GUI-Lib/src/main/java/de/muenchen/vaadin/guilib/components/ConfirmationWindow.java
@@ -1,16 +1,12 @@
-package de.muenchen.vaadin.guilib.components.buttons;
+package de.muenchen.vaadin.guilib.components;
 
-import com.vaadin.ui.Alignment;
-import com.vaadin.ui.Button;
-import com.vaadin.ui.HorizontalLayout;
-import com.vaadin.ui.Label;
-import com.vaadin.ui.VerticalLayout;
-import com.vaadin.ui.Window;
+import com.vaadin.ui.*;
 import com.vaadin.ui.themes.ValoTheme;
 import de.muenchen.vaadin.demo.i18nservice.I18nPaths;
 import de.muenchen.vaadin.demo.i18nservice.buttons.Action;
 import de.muenchen.vaadin.demo.i18nservice.buttons.SimpleAction;
 import de.muenchen.vaadin.guilib.BaseUI;
+import de.muenchen.vaadin.guilib.components.buttons.ActionButton;
 
 import java.util.ArrayList;
 import java.util.List;
